feat: Modularize Java writers and decouple Loader via reflection
