+++ hugegraph-loader-custom/src/main/java/org/apache/hugegraph/loader/writer/custom/NetworkXWriter.java
@@ -1,4 +1,4 @@
-package org.apache.hugegraph.loader.writer;  
+package org.apache.hugegraph.loader.writer.custom;  
   
 import java.io.BufferedReader;  
 import java.io.IOException;  
@@ -8,7 +8,8 @@ import java.nio.file.Paths;
 import java.util.*;  
 import java.util.concurrent.ConcurrentHashMap;  
   
-import org.apache.hugegraph.loader.builder.Record;  
+import org.apache.hugegraph.loader.builder.Record;
+import org.apache.hugegraph.loader.writer.Writer;  
 import org.apache.hugegraph.structure.graph.Edge;  
 import org.apache.hugegraph.structure.graph.Vertex;
