+++ pom.xml
@@ -91,6 +91,7 @@
     <modules>
         <module>hugegraph-client</module>
         <module>hugegraph-loader</module>
+        <module>hugegraph-loader-custom</module>
         <module>hugegraph-spark-connector</module>
         <module>hugegraph-tools</module>
         <module>hugegraph-hubble</module>
